% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.dataDescription.features.categorical.values}
\alias{Analyze.dataDescription.features.categorical.values}
\title{Analyze.dataDescription.features.categorical.values Object}
\usage{
Analyze.dataDescription.features.categorical.values()
}
\value{
Analyze.dataDescription.features.categorical.values object
}
\description{
Analyze.dataDescription.features.categorical.values Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}},
  \code{\link{Analyze}}
}

