% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.dataDescription.features}
\alias{Analyze.dataDescription.features}
\title{Analyze.dataDescription.features Object}
\usage{

  Analyze.dataDescription.features(Analyze.dataDescription.features.categorical = NULL,
  Analyze.dataDescription.features.categorical.values = NULL,
  Analyze.dataDescription.features.numeric = NULL,
  Analyze.dataDescription.features.text = NULL)
}
\arguments{
\item{Analyze.dataDescription.features.categorical}{The \link{Analyze.dataDescription.features.categorical} object or list of objects}

\item{Analyze.dataDescription.features.categorical.values}{The \link{Analyze.dataDescription.features.categorical.values} object or list of objects}

\item{Analyze.dataDescription.features.numeric}{The \link{Analyze.dataDescription.features.numeric} object or list of objects}

\item{Analyze.dataDescription.features.text}{The \link{Analyze.dataDescription.features.text} object or list of objects}
}
\value{
Analyze.dataDescription.features object
}
\description{
Analyze.dataDescription.features Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}},
  \code{\link{Analyze}}
}

