% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{hostedmodels.predict}
\alias{hostedmodels.predict}
\title{Submit input and request an output against a hosted model.}
\usage{
hostedmodels.predict(Input, project, hostedModelName)
}
\arguments{
\item{Input}{The \link{Input} object to pass to this method}

\item{project}{The project associated with the model}

\item{hostedModelName}{The name of a hosted model}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/prediction
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/prediction)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/prediction/docs/developer-guide}{Google Documentation}

Other Input functions: \code{\link{Input.input}},
  \code{\link{Input}}, \code{\link{trainedmodels.predict}}
}
