% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Input}
\alias{Input}
\title{Input Object}
\usage{
Input(Input.input = NULL, input = NULL)
}
\arguments{
\item{Input.input}{The \link{Input.input} object or list of objects}

\item{input}{Input to the model for a prediction}
}
\value{
Input object
}
\description{
Input Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Input functions: \code{\link{Input.input}},
  \code{\link{hostedmodels.predict}},
  \code{\link{trainedmodels.predict}}
}

