% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Insert}
\alias{Insert}
\title{Insert Object}
\usage{
Insert(Insert.trainingInstances = NULL, id = NULL, modelType = NULL,
  sourceModel = NULL, storageDataLocation = NULL,
  storagePMMLLocation = NULL, storagePMMLModelLocation = NULL,
  trainingInstances = NULL, utility = NULL)
}
\arguments{
\item{Insert.trainingInstances}{The \link{Insert.trainingInstances} object or list of objects}

\item{id}{The unique name for the predictive model}

\item{modelType}{Type of predictive model (classification or regression)}

\item{sourceModel}{The Id of the model to be copied over}

\item{storageDataLocation}{Google storage location of the training data file}

\item{storagePMMLLocation}{Google storage location of the preprocessing pmml file}

\item{storagePMMLModelLocation}{Google storage location of the pmml model file}

\item{trainingInstances}{Instances to train model on}

\item{utility}{A class weighting function, which allows the importance weights for class labels to be specified (Categorical models only)}
}
\value{
Insert object
}
\description{
Insert Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Insert functions: \code{\link{Insert.trainingInstances}},
  \code{\link{trainedmodels.insert}}
}

