% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Insert2.modelInfo}
\alias{Insert2.modelInfo}
\title{Insert2.modelInfo Object}
\usage{
Insert2.modelInfo(classWeightedAccuracy = NULL,
  classificationAccuracy = NULL, meanSquaredError = NULL,
  modelType = NULL, numberInstances = NULL, numberLabels = NULL)
}
\arguments{
\item{classWeightedAccuracy}{Estimated accuracy of model taking utility weights into account (Categorical models only)}

\item{classificationAccuracy}{A number between 0}

\item{meanSquaredError}{An estimated mean squared error}

\item{modelType}{Type of predictive model (CLASSIFICATION or REGRESSION)}

\item{numberInstances}{Number of valid data instances used in the trained model}

\item{numberLabels}{Number of class labels in the trained model (Categorical models only)}
}
\value{
Insert2.modelInfo object
}
\description{
Insert2.modelInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Model metadata.
}
\seealso{
Other Insert2 functions: \code{\link{Insert2}}
}
