% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Output.outputMulti}
\alias{Output.outputMulti}
\title{Output.outputMulti Object}
\usage{
Output.outputMulti()
}
\value{
Output.outputMulti object
}
\description{
Output.outputMulti Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Output functions: \code{\link{Output}}
}

